opti: mass in float as double is not needed

+++ include/components/mass.hpp
@@ -2,7 +2,7 @@
 
 namespace components {
     struct Mass {
-            double mantissa;
+            float mantissa;
             int exponent;
     };
 } // namespace components
